use a pointer to KX_GameObject instead of Object
This could be useful to identify calls if we continue to use the cache.

The main issue is that all calls are not added with
DRW_shgroup_call_object_add. When the calls are added with
DRW_shgroup_call_object_add, we can add a pointer to Object, but if the
calls are added with DRW_shgroup_call_add, or another function,
the DRWCall Object field is empty.

So we can use the cache only for calls added with
DRW_shgroup_call_object_add.

So the solution consisting to use the cache for KX_GameObject API is not
really suitable unless we modify eevee's sources...

But I'll maybe continue the stuff with that to see what we can do...