temp fix for end object
needs to be checked by brains and debuggers

Explainations:

To remove the rendered KX_GameObject, I tried to reconstruct all its
shgroups without all its DRWCalls (the KX_GameObject we want to remove
DRWCalls)

But when the shgroup is reconstructed, DRW_game_shgroup_call_add is
called for each objects/calls (other than the object we want to remove)
of the shgroup.

When we call DRW_game_shgroup_call_add, call->culled is set to false.

But we want the objects in inactive layers to be culled.

So Once the shgroup is reconstructed, we have to cull again all the objects in
scene inactive objects list...

This is not elegant at all, and very tricky, but waiting for a better
solution, I leave it like that