Add curves to scene.objects list
As we use eevee render, curves are rendered. Last time I already added
material batches to curves but as the curves were not converted as
KX_GameObject, they cannot move with game logic.

In this commit:

- As for lighprobe, we can convert curve as simple empty objects so
it can be transformed with gamelogic (animation, motion...) (e.g: some
trees are sometimes made with curves in some scenes I tested)

- As we convert curve as empty, it has no bounding box (gameobject
bounding box), so we don't apply culling to it.

maybe in blender code, curve have a BKE_object_bounding_box. If we'd
continue to work on bge for 2.8, I guess we should use
BKE_object_bounding_boxes instead of ours.

Well, I think my work will remain as experimental stuff anyway, and that
BF will switch directly to interactive mode (The more I work on eevee,
the more I think interactive mode is the right way to go, but as I don't
know how to deal with depsgraph, new UI for logic, bpy..
I'll maybe continue a bit the experimentations in this branch)