restore Level of detail (in bge) + fix shadows when replace mesh + try to make the shadows code more consistent with material code
oops: some issues to fix to replace shadows

Note: I don't like that but I had to duplicate gameobj API in
RAS_MeshObject. We should put all the code in KX_GameObject because
seral types of objects have Gwn_Batches (display arrays), not only
meshes: CURVES, FONTS....

But this will be a boring stuff to replace/remove RAS_MeshObject.

Ideally, the design I think about is to have in KX_GameObject:

1) Rendered Object
2) Logic Object
3) Physics object

everything in the same class, but with balises to make the code readable